refactor: Move ToSchema macro support traits to schema::derive_macro_support mod
